# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "更好的MAS游戏社区版",
    "version": "0.1.0"
  },
  "paths": {
    "/api/session": {
      "get": {
        "tags": [
          "Session"
        ],
        "summary": "Session",
        "operationId": "getSession",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Session"
        ],
        "summary": "Login Session",
        "operationId": "loginSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionLoginIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SessionLoginIn": {
        "properties": {
          "password": {
            "type": "string",
            "maxLength": 512,
            "minLength": 1,
            "format": "password",
            "title": "Password",
            "writeOnly": true
          }
        },
        "type": "object",
        "required": [
          "password"
        ],
        "title": "SessionLoginIn"
      },
      "SessionOut": {
        "properties": {
          "code": {
            "type": "integer",
            "title": "Code",
            "description": "状态码",
            "default": 200
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "操作状态",
            "default": "success"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "操作消息",
            "default": "操作成功"
          },
          "key": {
            "type": "string",
            "title": "Key",
            "default": ""
          },
          "loginRequired": {
            "type": "boolean",
            "title": "Loginrequired",
            "default": false
          }
        },
        "type": "object",
        "title": "SessionOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}